Refactor aws utils tests.
